+++ src/main/java/kristiania/no/jdbc/AbstractDao.java
@@ -1,9 +1,10 @@
 package kristiania.no.jdbc;
 
-import kristiania.no.jdbc.survey.Survey;
-
 import javax.sql.DataSource;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -57,7 +58,7 @@ public abstract class AbstractDao<T> {
     }
 
 
-    public void delete(int id, String sql) throws SQLException {
+    public void delete(long id, String sql) throws SQLException {
         try (Connection connection = dataSource.getConnection()) {
             try (PreparedStatement statement = connection.prepareStatement(sql)) {
                 statement.setLong(1, id);
